+++ Juego/src/Recolectable.hpp
@@ -16,7 +16,7 @@ class Recolectable : public INobjetos , public INdrawable //multiple herencia a 
         ~Recolectable();
         
         //drawable metodos
-        void setPosiciones(int nx,int ny,int nz);
+        void setPosiciones(float nx,float ny,float nz);
         void setID(int);
         int getID();
 
@@ -26,7 +26,7 @@ class Recolectable : public INobjetos , public INdrawable //multiple herencia a 
         float getRX();
         float getRY();
         float getRZ();
-        
+
         int getAtaque();
         const char* getNombre();
         const char* getObjeto();
